drivers: net: rswitch: Remove duplicate code
Move shared code from flower and matchall filters to common file.
