+++ drivers/net/ethernet/renesas/Makefile
@@ -9,7 +9,7 @@ ravb-objs := ravb_main.o ravb_ptp.o
 
 obj-$(CONFIG_RAVB) += ravb.o
 
-obj-$(CONFIG_RENESAS_ETHER_SWITCH) += rswitch.o rswitch_tc_u32.o rswitch_tc_flower.o rswitch_tc_matchall.o
+obj-$(CONFIG_RENESAS_ETHER_SWITCH) += rswitch.o rswitch_tc_u32.o rswitch_tc_flower.o rswitch_tc_matchall.o rswitch_tc_common.o
 
 rtsn_drv-objs := rtsn.o rtsn_ptp.o
