+++ drivers/net/ethernet/renesas/rswitch_tc_flower.c
@@ -7,6 +7,7 @@
 
 #include "rswitch.h"
 #include "rswitch_tc_filters.h"
+#include "rswitch_tc_common.h"
 
 static int rswitch_tc_flower_validate_match(struct flow_rule *rule)
 {
@@ -24,102 +25,6 @@ static int rswitch_tc_flower_validate_match(struct flow_rule *rule)
 		  )
 	    ) {
 		return -EOPNOTSUPP;
-	}
-
-	return 0;
-}
-
-static int rswitch_tc_flower_validate_action(struct rswitch_device *rdev,
-				struct flow_rule *rule)
-{
-	struct flow_action_entry *act;
-	struct flow_action *actions = &rule->action;
-	int i;
-	bool redirect = false, dmac_change = false, vlan_change = false;
-
-	flow_action_for_each(i, act, actions) {
-		switch (act->id) {
-		case FLOW_ACTION_DROP:
-			if (!flow_offload_has_one_action(actions)) {
-				pr_err("Other actions with DROP is not supported\n");
-				return -EOPNOTSUPP;
-			}
-			break;
-		case FLOW_ACTION_REDIRECT:
-			if (!ndev_is_rswitch_dev(act->dev, rdev->priv)) {
-				pr_err("Can not redirect to not R-Switch dev!\n");
-				return -EOPNOTSUPP;
-			}
-			redirect = true;
-			break;
-		case FLOW_ACTION_MANGLE:
-			if (act->mangle.htype != FLOW_ACT_MANGLE_HDR_TYPE_ETH) {
-				pr_err("Only dst MAC change is supported for mangle\n");
-				return -EOPNOTSUPP;
-			}
-			dmac_change = true;
-			break;
-		case FLOW_ACTION_VLAN_MANGLE:
-			if (be16_to_cpu(act->vlan.proto) != ETH_P_8021Q) {
-				pr_err("Unsupported VLAN proto for offload!\n");
-				return -EOPNOTSUPP;
-			}
-
-			vlan_change = true;
-			break;
-		default:
-			pr_err("Unsupported for offload action id = %d\n", act->id);
-			return -EOPNOTSUPP;
-		}
-	}
-
-	if (dmac_change && !redirect) {
-		pr_err("dst MAC change is supported only with redirect\n");
-		return -EOPNOTSUPP;
-	}
-
-	if (vlan_change & !redirect) {
-		pr_err("VLAN mangle is supported only with redirect\n");
-		return -EOPNOTSUPP;
-	}
-
-	return 0;
-}
-
-static int rswitch_tc_flower_setup_drop_action(struct rswitch_tc_filter *f)
-{
-	f->param.slv = 0x3F;
-	/* Explicitly zeroing parameters for drop */
-	f->param.dv = 0;
-	f->param.csd = 0;
-
-	return 0;
-}
-
-static int rswitch_tc_flower_setup_redirect_action(struct rswitch_tc_filter *f)
-{
-	struct rswitch_device *rdev = f->rdev;
-
-	f->param.slv = BIT(rdev->port);
-	f->param.dv = BIT(f->target_rdev->port);
-	f->param.csd = 0;
-
-	if (f->action & (ACTION_CHANGE_DMAC | ACTION_VLAN_CHANGE)) {
-		f->param.l23_info.priv = rdev->priv;
-		f->param.l23_info.routing_number = rswitch_rn_get(rdev->priv);
-		f->param.l23_info.routing_port_valid = BIT(rdev->port) | BIT(f->target_rdev->port);
-
-		if (f->action & ACTION_CHANGE_DMAC) {
-			ether_addr_copy(f->param.l23_info.dst_mac, f->dmac);
-			f->param.l23_info.update_dst_mac = true;
-		}
-
-		if (f->action & ACTION_VLAN_CHANGE) {
-			f->param.l23_info.update_ctag_vlan_id = true;
-			f->param.l23_info.update_ctag_vlan_prio = true;
-			f->param.l23_info.vlan_id = f->vlan_id;
-			f->param.l23_info.vlan_prio = f->vlan_prio;
-		}
 	}
 
 	return 0;
@@ -391,56 +296,6 @@ static int rswitch_tc_flower_setup_match(struct rswitch_tc_filter *f,
 	return 0;
 }
 
-static int rswitch_tc_flower_setup_action(struct rswitch_tc_filter *f,
-				struct flow_rule *rule)
-{
-	struct flow_action_entry *act;
-	struct flow_action *actions = &rule->action;
-	int i;
-
-	flow_action_for_each(i, act, actions) {
-		switch (act->id) {
-		case FLOW_ACTION_DROP:
-			f->action = ACTION_DROP;
-			break;
-		case FLOW_ACTION_REDIRECT:
-			f->action |= ACTION_MIRRED_REDIRECT;
-			f->target_rdev = netdev_priv(act->dev);
-			break;
-		case FLOW_ACTION_MANGLE:
-			/*
-			 * The only FLOW_ACT_MANGLE_HDR_TYPE_ETH is supported,
-			 * sanitized by rswitch_tc_flower_validate_action().
-			 */
-			f->action |= ACTION_CHANGE_DMAC;
-			rswitch_parse_pedit(f, act);
-			break;
-		case FLOW_ACTION_VLAN_MANGLE:
-			f->action |= ACTION_VLAN_CHANGE;
-			f->vlan_id = act->vlan.vid;
-			f->vlan_prio = act->vlan.prio;
-			break;
-		default:
-			/*
-			 * Should not come here, such action will be dropped by
-			 * rswitch_tc_flower_validate_action().
-			 */
-			pr_err("Unsupported action for offload!\n");
-			return -EOPNOTSUPP;
-		}
-	}
-
-	if (f->action & ACTION_DROP) {
-		return rswitch_tc_flower_setup_drop_action(f);
-	}
-
-	if (f->action & ACTION_MIRRED_REDIRECT) {
-		return rswitch_tc_flower_setup_redirect_action(f);
-	}
-
-	return -EOPNOTSUPP;
-}
-
 static int rswitch_tc_flower_replace(struct net_device *dev,
 				struct flow_cls_offload *cls_flower)
 {
@@ -451,7 +306,7 @@ static int rswitch_tc_flower_replace(struct net_device *dev,
 	int rc;
 
 	if (rswitch_tc_flower_validate_match(rule) ||
-		rswitch_tc_flower_validate_action(rdev, rule)) {
+		rswitch_tc_validate_flow_action(rdev, rule)) {
 		return -EOPNOTSUPP;
 	}
 
@@ -471,7 +326,7 @@ static int rswitch_tc_flower_replace(struct net_device *dev,
 	if (rc)
 		goto free;
 
-	rc = rswitch_tc_flower_setup_action(f, rule);
+	rc = rswitch_tc_setup_flow_action(f, rule);
 	if (rc)
 		goto put_pf;
